more refactoring to pull apart presenter
