+++ source/Sailfish/Analysis/ITestComputer.cs
@@ -3,7 +3,7 @@ using Sailfish.Contracts.Public;
 
 namespace Sailfish.Analysis;
 
-internal interface ITestComputer
+public interface ITestComputer
 {
     List<TestCaseResults> ComputeTest(TestData beforeTestData, TestData afterTestData, TestSettings settings);
 }
